Update Board.cs
Added missing RPi3B revision number as well as detection of 3A+, 3B+, and CM3+.  These will be detected as their non-plus versions but should work fine.  If there is a reason this doesn't work, adding a new model for 3A+/3B+ and CM3+ would be required.  I have no hardware to confirm any of these are correct and function as expected.